+++ ninja-core/src/main/java/ninja/Router.java
@@ -43,6 +43,9 @@ public interface Router {
      * @param clazz The controllerClass e.g. ApplicationController.class
      * @param methodName the methodName of the class e.g. "index"
      * @return The final url (without server, and without any prefixes)
+     * @deprecated Reverse routing in the Router is not validated and does not
+     *      URL-escape path or query parameters. Use <code>ninja.ReverseRouter</code>
+     *      to build your reverse routes.
      */
     public String getReverseRoute(Class<?> clazz, String methodName);
     
@@ -61,6 +64,9 @@ public interface Router {
      *          or simply use a String. If the raw uri does not contain the placeholders
      *          they will be added as query parameters ?key=value&key2=value2 and so on
      * @return The final url (without server, and without any prefixes)
+     * @deprecated Reverse routing in the Router is not validated and does not
+     *      URL-escape path or query parameters. Use <code>ninja.ReverseRouter</code>
+     *      to build your reverse routes.
      */
     public String getReverseRoute(Class<?> clazz, String methodName, Map<String, Object> parameterMap);
 
@@ -79,6 +85,9 @@ public interface Router {
      *          or simply use a String. If the raw uri does not contain the placeholders
      *          they will be added as query parameters ?key=value&key2=value2 and so on
      * @return The final url (without server, and without any prefixes)
+     * @deprecated Reverse routing in the Router is not validated and does not
+     *      URL-escape path or query parameters. Use <code>ninja.ReverseRouter</code>
+     *      to build your reverse routes.
      */
     public String getReverseRoute(Class<?> clazz, String methodName, Object ... parameterMap);
     
@@ -98,18 +107,40 @@ public interface Router {
      *          or simply use a String. If the raw uri does not contain the placeholders
      *          they will be added as query parameters ?key=value&key2=value2 and so on
      * @return The final url (without server, and without any prefixes)
+     * @deprecated Reverse routing in the Router is not validated and does not
+     *      URL-escape path or query parameters. Use <code>ninja.ReverseRouter</code>
+     *      to build your reverse routes.
      */
     public String getReverseRoute(Class<?> controllerClass,
                                  String controllerMethodName,
                                  Optional<Map<String, Object>> parameterMap);
         
-    
+    /**
+     * @deprecated Reverse routing in the Router is not validated and does not
+     *      URL-escape path or query parameters. Use <code>ninja.ReverseRouter</code>
+     *      to build your reverse routes.
+     */
     public String getReverseRoute(MethodReference controllerMethodRef);
     
+    /**
+     * @deprecated Reverse routing in the Router is not validated and does not
+     *      URL-escape path or query parameters. Use <code>ninja.ReverseRouter</code>
+     *      to build your reverse routes.
+     */
     public String getReverseRoute(MethodReference controllerMethodRef, Map<String, Object> parameterMap);
     
+    /**
+     * @deprecated Reverse routing in the Router is not validated and does not
+     *      URL-escape path or query parameters. Use <code>ninja.ReverseRouter</code>
+     *      to build your reverse routes.
+     */
     public String getReverseRoute(MethodReference controllerMethodRef, Object ... parameterMap);
     
+    /**
+     * @deprecated Reverse routing in the Router is not validated and does not
+     *      URL-escape path or query parameters. Use <code>ninja.ReverseRouter</code>
+     *      to build your reverse routes.
+     */
     public String getReverseRoute(MethodReference controllerMethodRef, Optional<Map<String, Object>> parameterMap);
     
     /**
@@ -123,6 +154,9 @@ public interface Router {
      * Returns the list of compiled routes.
      */
     public List<Route> getRoutes();
+    
+    public Optional<Route> getRouteForControllerClassAndMethod(
+        Class<?> controllerClass, String controllerMethodName);
 
     // /////////////////////////////////////////////////////////////////////////
     // convenience methods to use the route in a DSL like way